giant code cleanup

+++ common/playacem/allrondism/core/handlers/SpawnHandler.java
@@ -1,9 +1,9 @@
 package playacem.allrondism.core.handlers;
 
-import playacem.allrondism.lib.BlockIDs;
 import net.minecraftforge.event.Event.Result;
 import net.minecraftforge.event.ForgeSubscribe;
 import net.minecraftforge.event.entity.living.LivingSpawnEvent.CheckSpawn;
+import playacem.allrondism.lib.BlockIDs;
 
 /**
  * Allrondism
@@ -18,14 +18,15 @@ public class SpawnHandler {
 
     @ForgeSubscribe
     public void onEntitySpawn(CheckSpawn event) {
-        int x = (int)Math.floor((double)event.x);
-        int y = (int)Math.floor((double)event.y);
-        int z = (int)Math.floor((double)event.z);
-        y--;//block below the spawn place of the mob
+
+        int x = (int) Math.floor(event.x);
+        int y = (int) Math.floor(event.y);
+        int z = (int) Math.floor(event.z);
+        y--;// block below the spawn place of the mob
 
         if (event.world.getBlockId(x, y, z) == BlockIDs.STORAGE_BLOCKS) {
             event.setResult(Result.DENY);
-        }else {
+        } else {
             event.setResult(Result.DEFAULT);
         }
     }
